Made some changes to the files. Adding my helper to the project.
Rosie, will change the name for this particular function, since, we are
creating a new character. May change some of the other functionality as
well. Rosie will be Perry the Parrot... Until I find a way to make
Rosie more general, maybe take in a name from the user, then parse that
to be in the code... not sure if that's dooable. But name can be
changed at least.